+++ techbot/src/main/kotlin/com/firelib/techbot/command/RsiBolingerCommand.kt
@@ -86,7 +86,7 @@ class RsiBolingerCommand : CommandHandler {
 
     override fun handle(cmd: Cmd, bot: Bot, update: Update) {
 
-        val userId = update.chatId().getId().toInt()
+        val userId = update.chatId().getId()
 
         val value = transaction {
             Settings.select { (Settings.user eq userId) and (Settings.name eq name) }.map { it[Settings.value] }.firstOrNull()
