+++ techbot/src/main/kotlin/com/firelib/techbot/command/TfHandler.kt
@@ -25,7 +25,7 @@ class TfHandler : CommandHandler {
         val timeFrame = cmd.opts["tf"]!!
         val fromUser = update.fromUser()
 
-        val uid = fromUser.id.toInt()
+        val uid = fromUser.id
 
         BotHelper.ensureExist(fromUser)
