+++ techbot/src/main/kotlin/com/firelib/techbot/persistence/Settings.kt
@@ -3,7 +3,7 @@ package com.firelib.techbot.persistence
 import org.jetbrains.exposed.sql.Table
 
 object Settings : Table() {
-    val user = Settings.integer("user_id")
+    val user = Settings.long("user_id")
     val name = varchar("name", 50)
     val value = varchar("settings", 500)
