+++ techbot/src/main/kotlin/com/firelib/techbot/persistence/Subscriptions.kt
@@ -4,7 +4,7 @@ import firelib.finam.FinamDownloader
 import org.jetbrains.exposed.dao.id.IntIdTable
 
 object Subscriptions : IntIdTable() {
-    val user = integer("user_id")
+    val user = long("user_id")
     val ticker = varchar("ticker", 10)
     val market = varchar("market", 10).default(FinamDownloader.FinamMarket.SHARES_MARKET.id)
